// src/integration/myairops/gen/crm.ts
export interface components {
    schemas: {
// ...
        ContactModel: {
            fullName?: null | string;
            isCrew?: boolean;
            addresses?: null | components["schemas"]["AddressModel"][];
            contactMethods?: null | components["schemas"]["ContactMethodModel"][];
            crew?: components["schemas"]["CrewMember"];
            /** Format: int32 */
            primaryAddressId?: null | number | string;
            primaryTelephone?: null | string;
            primaryEmail?: null | string;
            primaryAddress?: null | string;
            idDocuments?: null | components["schemas"]["IdDocumentModel"][];
            executiveAssistants?: null | components["schemas"]["ContactExecutiveAssistantModel"][];
            executives?: null | components["schemas"]["ContactExecutiveAssistantModel"][];
            createdBy?: null | string;
            /** Format: date-time */
            createdDate?: null | string;
            modifiedByUser?: null | string;
            /** Format: date-time */
            lastModifiedDate?: null | string;
            isHidden?: boolean;
            contactNotes?: null | components["schemas"]["ContactNoteModel"][];
            passengerNotes?: null | components["schemas"]["PassengerNoteModel"][];
            /** Format: int32 */
            id?: number | string;
            /** @description This property will validate on the name field */
            title?: components["schemas"]["TitleModel"];
            firstNames?: null | string;
            lastName?: null | string;
            preferredName?: null | string;
            /** @description This property will validate on the name field */
            gender?: components["schemas"]["GenderModel"];
            /** Format: date-time */
            dateOfBirth?: null | string;
            isEmployee?: boolean;
            employeeId?: null | string;
            isPassenger?: boolean;
            passenger?: components["schemas"]["PassengerModel"];
            position?: null | string;
            honours?: null | string;
            salutation?: null | string;
            /**
             * Format: double
             * @description Weight is displayed in kilograms
             */
            weight?: null | number | string;
            tags?: null | string[];
            controlEmployee?: boolean;
            enabledForPersonalUse?: boolean;
            placeOfBirth?: null | string;
            areaOfBirth?: null | string;
            /** @description This property will validate on the ISO3 field */
            countryOfBirth?: components["schemas"]["CountrySummary"];
            externalReference?: null | string;
            /** Format: int32 */
            defaultSalesSectorTypeRegulationTypeId?: null | number | string;
            defaultCostCentre?: null | string;
            /** Format: int32 */
            sequenceNumber?: null | number | string;
            isTripApprover?: boolean;
            isTripRequestor?: boolean;
            isBillingContact?: boolean;
            /** @description Subject to clause 16a of the Securities Exchange Act 1934 */
            subjectToSEA1934?: boolean;
            knownTravellerNumber?: null | string;
            usRedressNumber?: null | string;
            securityConsideration?: boolean;
        };
// ...
        IdDocumentModel: {
            /** Format: int32 */
            id?: number | string;
            description?: null | string;
            /** @description This property will validate on the name field */
            idDocumentType?: components["schemas"]["IdDocumentTypeSummary"];
            firstName?: null | string;
            middleName?: null | string;
            lastName?: null | string;
            number?: null | string;
            placeOfBirth?: null | string;
            /** Format: date-time */
            issueDate?: null | string;
            /** Format: date-time */
            expiryDate?: null | string;
            /** @description This property will validate on the ISO3 field */
            issuingCountry?: components["schemas"]["CountrySummary"];
            /** @description This property will validate on the ISO3 field */
            residenceCountry?: components["schemas"]["CountrySummary"];
            /** @description This property will validate on the ISO3 field */
            citizenshipCountry?: components["schemas"]["CountrySummary"];
            /** @description This property will validate on the name field */
            nationality?: components["schemas"]["NationalitySummary"];
            visas?: null | components["schemas"]["VisaModel"][];
            isPreferred?: boolean;
            isHidden?: boolean;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}